Moved JSON-related stuff to JSONPipeline
