+++ script-server/src/main/kotlin/org/geobon/pipeline/Pipeline.kt
@@ -136,7 +136,7 @@ open class Pipeline (
         return "Pipeline(${debugName})"
     }
 
-    private fun initRoot() {
+    protected fun initRoot() {
         linkInputs()
         val errors = validateGraph()
         if (errors.isNotEmpty()) {
@@ -175,180 +175,6 @@ open class Pipeline (
             return miniPipeline
         }
 
-        fun createRootPipeline(serverContext: ServerContext, relPath: String, inputsJSON: String? = null) =
-            createRootPipeline(serverContext, File(pipelinesRoot, relPath), inputsJSON)
-
-        fun createRootPipeline(serverContext: ServerContext, descriptionFile: File, inputsJSON: String? = null): Pipeline {
-            return createFromFile(
-                serverContext,
-                StepId("", ""),
-                descriptionFile,
-                inputsJSON
-            ).apply {
-                initRoot()
-            }
-        }
-
-        fun createRootPipeline(
-            serverContext: ServerContext, debugName: String, pipelineJSON: JSONObject,
-            inputsJSON: JSONObject
-        ): Pipeline {
-            return createFromJSON(
-                serverContext,
-                StepId("", ""),
-                debugName,
-                pipelineJSON,
-                inputsJSON
-            ).apply {
-                initRoot()
-            }
-        }
-
-        private fun createFromFile(
-            serverContext: ServerContext, stepId: StepId, relPath: String,
-            inputsJSON: String? = null
-        ): Pipeline =
-            createFromFile(serverContext, stepId, File(pipelinesRoot, relPath), inputsJSON)
-
-        private fun createFromFile(
-            serverContext: ServerContext, stepId: StepId, descriptionFile: File,
-            inputsJSON: String? = null
-        ): Pipeline =
-            createFromJSON(
-                serverContext,
-                stepId,
-                descriptionFile.relativeTo(pipelinesRoot.parentFile).path,
-                JSONObject(descriptionFile.readText()),
-                inputsJSON?.let { JSONObject(inputsJSON) } ?: JSONObject()
-            )
-
-        private fun createFromJSON(
-            serverContext: ServerContext, stepId: StepId, debugName: String, pipelineJSON: JSONObject,
-            inputsJSON: JSONObject
-        ): Pipeline {
-            val logger = LoggerFactory.getLogger(debugName)
-
-            val constants = mutableMapOf<String, ConstantPipe>()
-            val outputIds = mutableListOf<String>()
-            val steps: MutableMap<String, IStep> = mutableMapOf()
-            val outputs: MutableMap<String, Output> = mutableMapOf()
-
-            // Load all nodes and classify them as steps, constants or pipeline outputs
-            pipelineJSON.getJSONArray(NODES_LIST).forEach { node ->
-                if (node is JSONObject) {
-                    val nodeId = node.getString(NODE__ID)
-                    when (node.getString(NODE__TYPE)) {
-                        NODE__TYPE_STEP -> {
-                            val script = node.getJSONObject(NODE__DATA)
-                                .getString(NODE__DATA__FILE)
-                            val scriptFile = script.replace('>', '/')
-
-                            val innerStepId = StepId(script, nodeId, stepId)
-                            steps[nodeId] = when {
-                                scriptFile.endsWith(".json") -> createFromFile(serverContext, innerStepId, scriptFile)
-
-                                // Instantiating kotlin "special steps".
-                                // Not done with reflection on purpose, since this could allow someone to instantiate any class,
-                                // resulting in a security breach.
-                                // TODO: This will be needed for openEO steps, so keeping this comment as an example:
-                                //scriptFile == "pipeline/AssignId.yml" -> AssignId(serverContext, innerStepId)
-
-                                // Regular script steps
-                                else -> ScriptStep(scriptFile, innerStepId, serverContext)
-                            }
-                        }
-
-                        NODE__TYPE_CONSTANT -> {
-                            val nodeData = node.getJSONObject(NODE__DATA)
-                            val type = nodeData.getString(NODE__DATA__TYPE)
-                            constants[nodeId] = createConstant(nodeId, nodeData, type, NODE__DATA__VALUE)
-                        }
-
-                        NODE__TYPE_USER_INPUT -> {
-                            val nodeData = node.getJSONObject(NODE__DATA)
-                            val type = nodeData.getString(NODE__DATA__TYPE)
-                            val userInputId = StepId("pipeline", nodeId, stepId)
-
-                            steps[nodeId] = UserInput(
-                                userInputId,
-                                type,
-                                pipelineJSON.optJSONObject(INPUTS)
-                                    ?.optJSONObject(userInputId.toString())
-                                    ?.optString(IO__LABEL, null)
-                            )
-                        }
-
-                        NODE__TYPE_OUTPUT -> outputIds.add(nodeId)
-                        else -> logger.warn("Ignoring node type ${node.getString(NODE__TYPE)}")
-                    }
-                } else {
-                    logger.warn("Unexpected object type under \"nodes\": ${node.javaClass}")
-                }
-            }
-
-            // Link steps & constants by reading the edges, and populate the pipelineOutputs variable
-            pipelineJSON.getJSONArray(EDGES_LIST).forEach { edge ->
-                if (edge is JSONObject) {
-                    // Find the source pipe
-                    val sourceId = edge.getString(EDGE__SOURCE_ID)
-                    val sourcePipe = constants[sourceId] ?: steps[sourceId]?.let { sourceStep ->
-                        val sourceOutput = edge.optString(EDGE__SOURCE_OUTPUT, Step.DEFAULT_OUT)
-                        sourceStep.outputs[sourceOutput]
-                            ?: throw Exception("Could not find output \"$sourceOutput\" in \"${sourceStep}\".\n" +
-                                    "Available outputs: ${sourceStep.outputs}")
-                    } ?: throw Exception("Could not find step with ID: $sourceId")
-
-                    // Find the target and connect them
-                    val targetId = edge.getString(EDGE__TARGET_ID)
-                    if (outputIds.contains(targetId)) {
-                        if (sourcePipe is Output) {
-                            val step = steps[sourceId]
-                            val outputId =
-                                if (step is Pipeline) IOId(step.id, sourcePipe.getId())
-                                else sourcePipe.getId()
-
-                            outputs[outputId.toString()] = sourcePipe
-                        } else {
-                            throw Exception("output in json not of Output type: $targetId")
-                        }
-                    } else {
-                        steps[targetId]?.let { step ->
-                            val targetInput = edge.getString(EDGE__TARGET_INPUT)
-                            step.inputs[targetInput] = step.inputs[targetInput].let {
-                                if (it == null) sourcePipe else AggregatePipe(listOf(it, sourcePipe))
-                            }
-                        } ?: logger.warn("Dangling edge: could not find target $targetId")
-                    }
-
-                } else {
-                    logger.warn("Unexpected object type under \"edges\": ${edge.javaClass}")
-                }
-            }
-
-            return Pipeline(
-                stepId,
-                debugName,
-                steps,
-                inputsToConstants(inputsJSON, pipelineJSON),
-                outputs
-            )
-        }
-
-        private fun inputsToConstants(inputsJSON: JSONObject, pipelineJSON: JSONObject): MutableMap<String, Pipe> {
-            val constants = mutableMapOf<String, Pipe>()
-            pipelineJSON.optJSONObject(INPUTS)?.let { inputsSpec ->
-                inputsJSON.keySet().forEach { key ->
-                    val inputSpec = inputsSpec.optJSONObject(key)
-                        ?: throw RuntimeException("Input received \"$key\" is not listed in pipeline inputs. Listed inputs are ${inputsSpec.keySet()}")
-                    val type = inputSpec.getString(IO__TYPE)
-
-                    constants[key] = createConstant(key, inputsJSON, type, key)
-                }
-            }
-
-            return constants
-        }
-
         private fun inputsToConstants(inputsString: String?, step: ScriptStep): MutableMap<String, Pipe> {
             if (inputsString == null)
                 return mutableMapOf()
@@ -366,7 +192,8 @@ open class Pipeline (
             return constants
         }
 
-        private fun createConstant(
+        @JvmStatic
+        protected fun createConstant(
             idForUser: String,
             obj: JSONObject,
             type: String,
@@ -420,34 +247,6 @@ open class Pipeline (
                     e.printStackTrace()
                     throw RuntimeException("Constant $idForUser has no value in JSON file.")
                 }
-            }
-        }
-
-        /**
-         * @param relativePath relative path to the JSON file
-         * @return the pipeline metadata as a JSONObject
-         * @see org.geobon.script.Description for return value structure
-         */
-        fun getPipelineDescription(relativePath: String): JSONObject {
-            val descriptionFile =
-                File(pipelinesRoot, relativePath)
-
-            if (descriptionFile.exists()) {
-                val descriptionJSON = JSONObject(descriptionFile.readText())
-                val metadataJSON = JSONObject()
-                metadataJSON.putOpt(INPUTS, descriptionJSON.get(INPUTS))
-                metadataJSON.putOpt(OUTPUTS, descriptionJSON.get(OUTPUTS))
-
-                descriptionJSON.optJSONObject(METADATA)?.let { metadata ->
-                    metadata.keys().forEach { key ->
-                        metadataJSON.putOpt(key, metadata.get(key))
-                    }
-                }
-
-                return metadataJSON
-
-            } else {
-                throw FileNotFoundException("$descriptionFile does not exist.")
             }
         }
     }
