+++ script-server/src/main/kotlin/org/geobon/server/plugins/Routing.kt
@@ -13,7 +13,7 @@ import org.geobon.k8s.K8sConnection
 import org.geobon.openeo.getOpenEODescription
 import org.geobon.pipeline.*
 import org.geobon.pipeline.Pipeline.Companion.createMiniPipelineFromScript
-import org.geobon.pipeline.Pipeline.Companion.createRootPipeline
+import org.geobon.pipeline.JSONPipeline.Companion.createRootPipeline
 import org.geobon.server.ServerContext
 import org.geobon.server.ServerContext.Companion.pipelinesRoot
 import org.geobon.server.ServerContext.Companion.scriptStubsRoot
@@ -165,7 +165,7 @@ fun Application.configureRouting() {
 
                     "pipeline" -> {
                         val jsonPath = descriptionPath.replace('>', '/')
-                        call.respondText(Pipeline.getPipelineDescription(jsonPath).toString(), ContentType.Application.Json)
+                        call.respondText(JSONPipeline.getPipelineDescription(jsonPath).toString(), ContentType.Application.Json)
                     }
                     "openEO" -> {
                         call.respondText(
